backend/backend-app/requirements.txt
Flask==2.1.1
Werkzeug==2.0.3  # Pinning a compatible version of Werkzeug
boto3==1.20.24
psycopg2-binary==2.9.3
flask-cors==3.0.10